Add check to fix regression with the file cleanup for builds

+++ lib/sproutcore/tools/build.rb
@@ -116,8 +116,10 @@ module SC
         paths_array = paths.split(",")
         paths_array.each do |entry|
           if not entry.include? "index.html"
-            puts "Deleting "+entry
-            File.delete(entry)
+            if File.exist? (entry)
+              puts "Deleting "+entry
+              File.delete(entry)
+            end
           end
         end
       end
